fix(ai): unify workout analysis schema scores and status vocabulary (CW-403)
